Creating and implementing module for making normal markup editable
Creating editable.js . It can take markup and make corresponding edit boxes for
any elements with the class '.editable'.
I expect there's a backbone module available that does this.
It felt like a good thing to make it myself as it's fairly simple, has helped me to understand the subject matter better and may save me time in future.