Making follow related backbone objects work
Took a performance optimising approach to fetching followers list.
In the absence of any user details stored in the Follow Service it seemed important to avoid
doing a separate user get request for each follower.

Working on the assumption that all users are loaded for the main list I added a method to
the add for each of these which copies the user into a 'followers' collection if appropriate.

If in future not all users were loaded then more data would need to redundantly sit on the follow service.
Second best would be for the app server to get a full list of followers from the service and use this to filter
its response.